Fixed validation bug in startMatch that prevented adding any match unless it's the first by moving the check outside the loop before starting a new match.
In Scoreboard.java edited the startMatch method to allow matches to start in the past for testing purposes

Added a check in updateScore to check if a match has started before updating the score.

Fix finishMatch method – replaced for loop with removeIf() to safely remove matches, ensured only ongoing matches are considered for removal, and improved error handling for non-existing matches.

Updated getSummary to sort matches correctly first by score and then start time the first being the most recent